Narrow Bx storage from f64 to f32 in orthonormalize()
- kernels.{h,cc}: orthonormalize()'s Bx parameter goes from double* to
  float*. All arithmetic (fill, T = Bo^T*Bx, project-subtract,
  normalize) stays in f64; only the per-row store narrows. Phase 2a's
  s_buf accumulator uses post-store (f32-rounded) values so the
  downstream normalize step sees the same magnitude that's in memory.
  compute_BoT_bx_col / project_subtract_and_norm take float*; the
  Phase 0 fill drops its f64 cvt (just stores f32).

- marsalgo.cc: MarsScratch::Impl::Bx is MatrixXf. eval()'s linear_dsse
  switches from a single `ybx = Bx.transpose() * y` matmul (Eigen
  mixed-scalar would require materializing an n*p f64 temp) to a
  per-column `ybx[j] = Bx.col(j).cast<double>().dot(y)` -- lazy cast,
  no temp. Hinge-loop `bx` pointer is now float*; the existing
  `double bx_k = bx[k[i]]` does the upcast at the load. Updated the
  scratch-block comment to reflect the change and the remaining Bo
  follow-up.

- tests/test_kernels.cc: existing four tests get MatrixXf Bx +
  .cast<double>() at assertion sites + tolerances at the f32 floor
  (1e-5 for orth / norm / vs-ref). New stress test
  OrthonormalizePrecisionStressBaseline locks in the precision floor
  under a mixed configuration (well-conditioned + DGKS-triggering +
  exactly-degenerate columns) so future changes are measured against
  it rather than eyeballed.

- tests/unittest.cc: MarsTest.DeltaSSE hinge dsse/N tolerances
  relaxed 1e-7/1e-8 -> 1e-6. The linear-path absolute-dsse tolerances
  stay at 1e-7. The hinge floor is higher because the inner sweep
  accumulates f32 noise through f/g over n iterations, while the
  linear case is bounded by the ybx[j] dot product alone.

Measured under linear_only=True, threads=-1:
  n=100k p=1200 epochs=6 (M=7): 934-953 -> 850-864 ms/epoch (~1.10x)

The bigger win is the per-thread footprint of the largest scratch
buffer (Bx) halving -- ~150 GB across 64 threads at the production
n=millions, M=400 scale. Bo -> f32 remains a separate follow-up; the
DGKS-cancellation risks compound across the M~400 appends and need
their own measurement.

+++ kernels.h
@@ -19,20 +19,27 @@ constexpr double DGKS_GATE_RATIO_SQ = 9.0;
  *  Interact a candidate regressor `x` with the basis columns selected by
  *  `mask`, orthonormalize against `Bo`, and normalize each column. For each
  *  j in [0, p):
- *      Bx[:,j]  = (double)(B[:, mask[j]] .* x)             // fill (f32 * f32 → f64)
- *      Bx[:,j] -= Bo * (Bo^T * Bx[:,j])                    // project out Bo's span
- *      s        = ||Bx[:,j]||^2
+ *      Bx[:,j]  = B[:, mask[j]] .* x                       // fill (f32 * f32 → f32 store)
+ *      Bx[:,j] -= Bo * (Bo^T * Bx[:,j])                    // project out Bo's span (f64 arith)
+ *      s        = ||Bx[:,j]||^2                            // f64 accumulator over post-store f32
  *      Bx[:,j] *= (s > tol) ? 1/sqrt(s + tol) : 0          // normalize, zero degenerate
+ *
+ *  Bx is stored as f32 to halve the per-thread scratch footprint; all
+ *  arithmetic (projection, dot products, norms) stays in f64. The
+ *  cancellation in (Bx -= Bo * Bo^T * Bx) is bounded by the DGKS retry, so
+ *  the rounded f32 storage keeps O(eps_f32) orthogonality against Bo --
+ *  see DGKS_GATE_RATIO_SQ below.
  *
  *  Layouts (caller's responsibility):
  *      B     : (n, *) col-major float;  col stride ldB.   Only columns mask[j] are read.
  *      x     : (n)    float
  *      mask  : (p)    int32, indexes into columns of B
  *      Bo    : (n, m) row-major double; row stride ldBo
- *      Bx    : (n, p) col-major double; col stride ldBx       [output]
+ *      Bx    : (n, p) col-major float;  col stride ldBx       [output]
  *      T     : (m, p) col-major double; col stride ldT        [workspace; overwritten]
  *      s_buf : (p)    double                                  [workspace; overwritten]
- *      tol   : tolerance for treating a column as degenerate
+ *      tol   : tolerance for treating a column as degenerate;
+ *              should be ~O(n * eps_f32^2) for the f32 storage floor.
  *
  *  Uses AVX2 + FMA in the inner loops when available, scalar fallback otherwise.
  *  No heap allocations.
@@ -47,7 +54,7 @@ void orthonormalize(
     const float  *x,
     const int    *mask,
     const double *Bo,   int ldBo,
-    double       *Bx,   int ldBx,
+    float        *Bx,   int ldBx,
     double       *T,    int ldT,
     double       *s_buf,
     double tol,
